xen/intrng: move isrc allocation to architecture (INTRNG version)
The isrc allocation has been moved to architecture code where
the decision to allocate or not can be made.

This means setting of several fields needs to be handled by architecture
code.  Modify the prototype of xen_arch_intr_alloc(), start passing the
variables needed, and finally set them on aarch64.

This also means releasing of isrcs also needs to be moved to
architecture code.

Differential Revision: https://reviews.freebsd.org/D31063