kern/intr: remove "irq" argument from intr_event_create()
Having the "irq" as an argument to intr_event_create() requires knowing
the value before calling.  This matches the approach of traditional x86
interrupts where the value was always known early.  Other architectures
and designs may not know the value until later.  As such it is better
not to be an argument and simply be set later.

Differential Revision: https://reviews.freebsd.org/D32504