kern/intr: have intr_event_destroy() return success on NULL event
A NULL event likely indicates the event was either already released or
never allocated.  In such case the event has already been cleaned up and
that is what the caller needs to know.  This gets called from cleanup
paths and cleanup should not be abandoned if the event is invalid.

Emit a warning message as this indicates inconsistent state.  There may
well simply be the precursor to a page-fault panic().

Differential Revision: https://reviews.freebsd.org/D38602