kern/bus: start initializing root bus earlier
This allows architecture to attach devices before normal bus probing.
This allows x86 to convert PICs to actual devices.

Additionally the TASKQ SWIs are initialized early.  There are also a few
SWIs during SOFTINTR.

Hopefully to be removed in the future.

Switch to static initialization for bus_data_devices as this removes
another extra bit from the initializer.