kern/intr: interpret pointer to intr_event as "source" value
All callers of `intr_event_init()` are passing a structure surrounding
the event structure.  As a result the pointers alias and a pointer to
the event can be converted back to the "source".

This makes the relationship between machine interrupt core and interrupt
events, similar to the relationship between PICs and machine interrupt
core.