kern/intr: add non-NULL event interface variant
Some callers are known to have already initialize the event before
calling these functions.  Other callers actively want to check locally
before calling these.  As such add variants which assume the event will
be non-NULL and ASSERT this being true.

PowerPC has a KASSERT() before `intr_event_handle()`, so blame the
original author if it is possible for the event to be NULL there.

Differential Revision: https://reviews.freebsd.org/D50835