x86/isa/clock: removing caching of ->pic_source_pending()
This amounts to a revert of f9acc6410e5.  This micro-optimization no
longer seems worthwhile.  More troublesome, pointers to functions don't
work with kernel object methods.  As such get rid of the caching of the
pointer and simply use the method.

f9acc6410e5 made the ->pic_source_pending hook optional.  This was
merely a consequence of the implementation approach.  At this point
though the hook is no longer commonly used, so keep the hook optional.